+++ src/automaton/tools/miner/miner.cc
@@ -23,7 +23,7 @@ namespace miner {
 
 std::string str_tohex(std::string s) {
   std::stringstream ss;
-  for (int i = 0; i < s.size(); i++) {
+  for (size_t i = 0; i < s.size(); i++) {
     ss << std::hex << std::uppercase << std::setw(2) << std::setfill('0') << (static_cast<int>(s[i]) & 0xff);
   }
   return ss.str();
